fix get PostgresURL
