+++ src/utils.ts
@@ -24,3 +24,16 @@ export const fetch = (url: string, init: RequestInit): Promise<Response> => {
     throw err;
   });
 }
+
+export const getPostgresURL = () => {
+  if (!process.env.POSTGRES_URL) {
+    return '';
+  }
+
+  const url = new URL(process.env.POSTGRES_URL);
+  if (url.hostname.includes('postgres.vercel-storage.com') && !url.searchParams.get('sslmode')) {
+    url.searchParams.set('sslmode');
+  }
+
+  return url.toString();
+}
